[None][fix] GVR Top-K: address PR review nits (coderabbitai)
Apply four coderabbitai-flagged tweaks:

1. ``gvr_topk_decode.py`` GvrParams docstring — replace ``σ`` / ``×``
   with ASCII (``sigma`` / ``x``) to silence ruff RUF002 on this
   tuning note.

2. ``gvr_topk_decode.py`` ``__call__`` — add a launch-time guard so the
   ``enable_smem_cache=True`` path bails out when the worst-case
   per-CTA slice exceeds the JIT-baked ``smem_cache_elems`` budget.
   Without this, ``load_slice_to_smem`` and the Phase 2 / Phase 3
   readers would silently overrun the fixed-size ``smem_input`` buffer
   on any caller that flips the flag without the matching size check.
   Wrapped in ``cutlass.const_expr(self.enable_smem_cache)`` so the
   default (cache disabled) path compiles to identical SASS.

3. ``run_gvr_topk.py`` ``gvr_topk_lb_prepare`` — raise ValueError when
   ``seq_lens.shape[0] > max_batch_size``. The LB prepare kernel
   hard-wires a block of ``max_batch_size`` threads and an
   ``order_row[max_batch_size]`` buffer; tail requests beyond that
   would never be classified into the long/short partition and the
   decode path's ``order_row`` lookup would return invalid row indices.

4. ``test_cute_dsl_gvr_topk_decode.py`` — drop redundant ``int()``
   wrapper around ``round()`` (Python 3 ``round(int)`` already returns
   int).

+++ tests/unittest/_torch/attention/sparse/test_cute_dsl_gvr_topk_decode.py
@@ -348,7 +348,7 @@ def test_lb_prepare_partition(B, ratio):
     """
     long_threshold = 64 * 1024
     torch.manual_seed(B * 1000 + int(ratio * 100))
-    n_long_expect = int(round(B * ratio))
+    n_long_expect = round(B * ratio)
     seq_lens = torch.empty(B, dtype=torch.int32, device="cuda")
     seq_lens[:n_long_expect] = long_threshold * 2
     seq_lens[n_long_expect:] = long_threshold // 2
